new templates for stat application, link to stats in the admin panel

+++ trunk/diamandas/stats/views.py
@@ -17,7 +17,7 @@ def entries(request):
 	"""
 	return unique entries in days
 	"""
-	if request.user.is_authenticated() and request.user.has_perm('stats.add_stat'):
+	if request.user.is_authenticated() and request.user.is_staff:
 		today = str(datetime.today())[:10]
 		tday = int(today[8:10])
 		results = []
@@ -39,10 +39,11 @@ def referers(request):
 	"""
 	return non SE/google referers
 	"""
-	if request.user.is_authenticated() and request.user.has_perm('stats.add_stat'):
+	if request.user.is_authenticated() and request.user.is_staff:
 		refs = Stat.objects.order_by('-id')
-		refs = refs.exclude(referer__icontains='rk.edu.pl')
+		refs = refs.exclude(referer__icontains=settings.SITE_KEY)
 		refs = refs.exclude(referer__icontains='google')
+		refs = refs.exclude(referer__icontains='yahoo')
 		refs = refs.exclude(referer__icontains='msn.com')
 		refs = refs.exclude(referer__icontains='onet.pl')
 		refs = refs.exclude(referer__icontains='netsprint.pl')
@@ -55,7 +56,7 @@ def google(request):
 	"""
 	return list of google keywords
 	"""
-	if request.user.is_authenticated() and request.user.has_perm('stats.add_stat'):
+	if request.user.is_authenticated() and request.user.is_staff:
 		refs = Stat.objects.order_by('-id').filter(referer__icontains='google').values('referer')
 		words = []
 		words_keys = {}
@@ -78,18 +79,12 @@ def google(request):
 		all_links = refs = float(Stat.objects.all().count())
 		google_pr = str((google_links/all_links)*100)[0:5]
 		
-		jaki = refs = float(Stat.objects.filter(referer__icontains='jakilinux').count())
-		jaki_pr = str((jaki/all_links)*100)[0:5]
-		
-		wykop = refs = float(Stat.objects.filter(referer__icontains='wykop').count())
-		wykop_pr = str((wykop/all_links)*100)[0:5]
-		
 		wiki = refs = float(Stat.objects.filter(referer__icontains='wikipedia').count())
 		wiki_pr = str((wiki/all_links)*100)[0:5]
 		
 		return render_to_response(
 			'stats/google.html',
-			{'refs': results, 'google': google_pr, 'jaki': jaki_pr, 'wykop': wykop_pr, 'wiki': wiki_pr},
+			{'refs': results, 'google': google_pr, 'wiki': wiki_pr},
 			context_instance=RequestContext(request))
 	return render_to_response('pages/bug.html', {'bug': _('You don\'t have the permissions to view this page')}, context_instance=RequestContext(request))
 
@@ -98,12 +93,16 @@ def mainpage(request):
 	"""
 	main page
 	"""
-	return render_to_response('stats/mainpage.html', context_instance=RequestContext(request))
+	if request.user.is_authenticated() and request.user.is_staff:
+		return render_to_response('stats/mainpage.html', context_instance=RequestContext(request))
+	else:
+		return render_to_response('pages/bug.html', {'bug': _('You don\'t have the permissions to view this page')}, context_instance=RequestContext(request))
 
 
 def delete_all(request):
 	"""
 	delete all data
 	"""
-	Stat.objects.all().delete()
+	if request.user.is_authenticated() and request.user.is_staff:
+		Stat.objects.all().delete()
 	return HttpResponseRedirect('/')
